This is an itermediate commit. Changed stdlib flag for clang to libstdc++.
The Floating values Dialog now has a FloatingValuesDialogData struct as
member. We use this struct to store the data that was selected by the
user. This way we get rid of our signal and no longer need any
SignalMappers for this. We will use lambdas to connect all of our
ClickableLabels to the dialog. This means way more refactoring to do.

DisplayArea is now dynamically building the Data display/control widget.

Signed-off-by: Christian Rapp <0x2a@posteo.org>