Code improvements and better handling of changed Device
Removed lots of constant references of primitive types as they are
unnecessary.
PowerSupplySCPI now has a member deviceName that stores the name the
device was given by the user in the device wizard. this is used to
determine ( among other things ) if the device has changed and the
controller has to create a new object.

Signed-off-by: Christian Rapp <0x2a@posteo.org>