Add Logging area to settings dialog
I have added a complete logging section to the settings dialog. Logging
is done with ealogger. You can choose the log files directory and the
minimum log level.

Added ealogger as CMake ExternalProject. This will build ealogger along
the main application as static library. Makes it easier for users that
compile the application from source.

The application checks when starting if there are more than 5 log files
and deletes them.

Make sure the cache and data location actually exist. If not we create
them.