Add log level elements to cbox
I have added the missing log level elements to the combobox in the
logging section.

Fixed out_of_range exception because I have removed the default path
from general_defaults as this was initialized to early and the directory
was not correct. The default directory is now a private member of
SettingsDialog like the default sqlite file.

Signed-off-by: Christian Rapp <0x2a@posteo.org>