Add Meyers Singleton for logger
The logger instance is now managed via a singleton pattern based on
Scott Meyer. This is the only way to provide designer based gui classes
with a logger object.

Added an option to the logging section to flush log files.

The logger is now initialized in main()

Started to replace qdebug calls with ealogger

Replaced public enums with strongly typed enum classes.