Preliminary docker / AppImage support
I have added a Dockerfile for a centos (6/7) based build system thats
linuxdeployqt to create an AppImage executable that should run on many
different linux distributions. Currently this works not very well as
there are dependency problems with libselinux. This dependency is
blacklisted by linuxdeployqt and not present on systems like archlinux.
As a result the appimage won't work there.

Signed-off-by: Christian Rapp <0x2a@posteo.org>